flask application for my website V1
This is the Flask application I deployed for my website. It only sends an automated email.
If you want to try it, you will need a Gmail account with a key that you must store in a .env file.
To install the virtual environment for the production environment, you will need to run these three commands:

python -m venv venv
./venv/Scripts/activate
pip install -r requirements.txt